requirements.txt
asgiref==3.9.1
dj-config-url==0.1.1
Django==5.2.4
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
